+++ admin/src/Field/CustomgetsField.php
@@ -43,24 +43,25 @@ class CustomgetsField extends ListField
 	protected function getOptions()
 	{
 		$db = Factory::getDBO();
-		$query = $db->getQuery(true);
-		$query->select($db->quoteName(array('a.id','a.name','a.gettype'),array('id','custom_get_name','type')));
-		$query->from($db->quoteName('#__componentbuilder_dynamic_get', 'a'));
-		$query->where($db->quoteName('a.published') . ' = 1');
+$query = $db->getQuery(true);
+$query->select($db->quoteName(array('a.id','a.name','a.gettype'),array('id','custom_get_name','type')));
+$query->from($db->quoteName('#__componentbuilder_dynamic_get', 'a'));
+$query->where($db->quoteName('a.published') . ' = 1');
 		$query->where('( '.$db->quoteName('a.gettype') . ' = 3 OR ' . $db->quoteName('a.gettype') . ' = 4 )');
 		$query->order('a.name ASC');
-		$db->setQuery((string)$query);
-		$items = $db->loadObjectList();
-		$options = array();
-		if ($items)
-		{
-			$model = ComponentbuilderHelper::getModel('dynamic_gets');
-			foreach($items as $item)
-			{
+$db->setQuery((string)$query);
+$items = $db->loadObjectList();
+$options = array();
+if ($items)
+{
+	$model = ComponentbuilderHelper::getModel('dynamic_gets');
+	foreach($items as $item)
+	{
 				$type = $model->selectionTranslation($item->type,'gettype');
-				$options[] = Html::_('select.option', $item->id, $item->custom_get_name . ' (' . Text::_($type) . ')' );
-			}
-		}
-		return $options;
+		$options[] = Html::_('select.option', $item->id, $item->custom_get_name . ' (' . Text::_($type) . ')' );
+	}
+}
+
+return $options;
 	}
 }